Level System 0.2

+++ Assets/Script/PlayerStatsController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -11,7 +12,7 @@ public class PlayerStatsController : MonoBehaviour
     public float xpFirstLevel = 100;
     public float difficultFactor = 1.5f;
 
-    private float xpNextLevel;
+    private static float xpNextLevel;
 
     // Start is called before the first frame update
     void Start()
@@ -20,24 +21,30 @@ public class PlayerStatsController : MonoBehaviour
         DontDestroyOnLoad(gameObject);
         Application.LoadLevel("GamePlay");
 
-        xpNextLevel = xpFirstLevel * (GetCurrentLevel() + 1) * difficultFactor;
-        
-        Debug.Log(xpNextLevel);
-        
-        PlayerPrefs.DeleteAll();
-        AddXp(10);
-        Debug.Log(GetCurrentXp());
     }
 
     // Update is called once per frame
     void Update()
     {
-        
+        if (Input.GetKeyDown(KeyCode.A))
+        {
+            AddXp(100);
+        }
+
+        if (Input.GetKeyDown(KeyCode.R))
+        {
+            PlayerPrefs.DeleteAll();
+        }
     }
 
     public static void AddXp(float xpAdd)
     {
         float newXp = (GetCurrentXp() + xpAdd) * PlayerStatsController.Instance.xpMultiply;
+        while (newXp >= GetNextXP())
+        {
+            newXp -= GetNextXP();
+            AddLevel();
+        }
         PlayerPrefs.SetFloat("currentXp", newXp);
     }
 
@@ -55,5 +62,17 @@ public class PlayerStatsController : MonoBehaviour
     {
         int newLevel = GetCurrentLevel()+1;
         PlayerPrefs.SetInt("currentLevel", newLevel);
+    }
+
+    public static float GetNextXP()
+    {
+        return PlayerStatsController.Instance.xpFirstLevel * (GetCurrentLevel() + 1) * PlayerStatsController.Instance.difficultFactor;
+    }
+
+    private void OnGUI()
+    {
+        GUI.Label(new Rect(0, 0, 150, 50), "Currente XP = " + GetCurrentXp());
+        GUI.Label(new Rect(0, 20, 150, 50), "Currente Level = " + GetCurrentLevel());
+        GUI.Label(new Rect(0, 40, 150, 50), "Currente Next XP = " + GetNextXP());
     }
 }
